refactor(visualization): split color popover into section components
The popover and folder overrides templates had grown too large to read.
The popover is now a thin composition of focused section components:

- cc-color-settings-header (metric name, gradient dot, reset thresholds)
- cc-color-range-section (threshold slider + distribution diagram)
- cc-gradient-mode-picker (segmented control)
- cc-color-bands-section (band rows with counts)
- cc-invert-reset-row (invert toggle + reset colors)
- cc-folder-override-row / cc-pin-folder-search (extracted from
  cc-folder-overrides)

Each section owns its service wiring, and the moved logic took its
tests along (debounced threshold commit, threshold reset, inversion
flags, reset keys, gradient mode dispatch). No behavior change.
