+++ visualization/app/codeCharta/features/metricsBar/components/colorSettingsPopover/colorSettingsPopover.component.html
@@ -3,138 +3,16 @@
     [anchorName]="anchorName()"
     [widthClass]="'max-w-[95vw] ' + (hasRangeSection() ? 'w-[420px]' : 'w-80')"
 >
-    <div class="flex items-center justify-between gap-2">
-        <div class="flex min-w-0 items-center gap-2">
-            <span class="inline-block h-3.5 w-3.5 shrink-0 rounded-full" [style.background]="headerSwatchGradient()"></span>
-            <span class="truncate text-sm">Color by <span class="font-semibold text-primary">{{ colorMetric() }}</span></span>
-        </div>
-        @if (hasRangeSection()) {
-            <button
-                class="btn btn-ghost btn-xs btn-square"
-                title="Reset slider thresholds to default"
-                aria-label="Reset thresholds"
-                (click)="resetColorRange()"
-            >
-                <i class="fa fa-undo"></i>
-            </button>
-        }
-    </div>
+    <cc-color-settings-header></cc-color-settings-header>
 
     @if (hasRangeSection()) {
-        <cc-metric-color-range-slider
-            class="block"
-            [minValue]="sliderValues().min"
-            [maxValue]="sliderValues().max"
-            [currentLeftValue]="sliderValues().from"
-            [currentRightValue]="sliderValues().to"
-            [leftColor]="sliderColors().leftColor"
-            [middleColor]="sliderColors().middleColor"
-            [rightColor]="sliderColors().rightColor"
-            [handleValueChange]="handleValueChange"
-            [sliderWidth]="220"
-        ></cc-metric-color-range-slider>
-
-        <div class="flex items-center justify-between gap-2">
-            <span class="text-xs uppercase tracking-wider opacity-60">Distribution</span>
-            <span class="truncate text-xs opacity-60">{{ colorMetric() }} × quantile %</span>
-        </div>
-        <cc-metric-color-range-diagram
-            [minValue]="sliderValues().min"
-            [maxValue]="sliderValues().max"
-            [currentLeftValue]="sliderValues().from"
-            [currentRightValue]="sliderValues().to"
-            [leftColor]="sliderColors().leftColor"
-            [middleColor]="sliderColors().middleColor"
-            [rightColor]="sliderColors().rightColor"
-            [colorMetric]="colorMetric()"
-            [values]="sliderValues().values"
-            [isAttributeDirectionInverted]="isAttributeDirectionInversed()"
-            [diagramWidth]="380"
-            [diagramHeight]="190"
-            [showAxisLabels]="false"
-        ></cc-metric-color-range-diagram>
-
-        <div class="flex flex-col gap-1">
-            <span class="text-xs uppercase tracking-wider opacity-60">Gradient Mode</span>
-            <div class="join w-full" role="radiogroup" aria-label="Gradient mode">
-                <input
-                    type="radio"
-                    [name]="popoverId() + '-gradient-mode'"
-                    class="join-item btn btn-sm flex-1"
-                    aria-label="Absolute"
-                    [checked]="colorMode() === 'absolute'"
-                    (change)="handleColorModeChange('absolute')"
-                />
-                <input
-                    type="radio"
-                    [name]="popoverId() + '-gradient-mode'"
-                    class="join-item btn btn-sm flex-1"
-                    aria-label="Focused"
-                    [checked]="colorMode() === 'focusedGradient'"
-                    (change)="handleColorModeChange('focusedGradient')"
-                />
-                <input
-                    type="radio"
-                    [name]="popoverId() + '-gradient-mode'"
-                    class="join-item btn btn-sm flex-1"
-                    aria-label="Weighted"
-                    [checked]="colorMode() === 'weightedGradient'"
-                    (change)="handleColorModeChange('weightedGradient')"
-                />
-                <input
-                    type="radio"
-                    [name]="popoverId() + '-gradient-mode'"
-                    class="join-item btn btn-sm flex-1"
-                    aria-label="Relative"
-                    [checked]="colorMode() === 'trueGradient'"
-                    (change)="handleColorModeChange('trueGradient')"
-                />
-            </div>
-        </div>
+        <cc-color-range-section></cc-color-range-section>
+        <cc-gradient-mode-picker [groupName]="popoverId() + '-gradient-mode'"></cc-gradient-mode-picker>
     }
 
-    <div class="flex flex-col gap-1">
-        <span class="text-xs uppercase tracking-wider opacity-60">Bands</span>
-        @if (!isDeltaState()) {
-            <cc-color-band-row mapColorFor="positive" [count]="colorCategoryCounts().positive"></cc-color-band-row>
-            <cc-color-band-row mapColorFor="neutral" [count]="colorCategoryCounts().neutral"></cc-color-band-row>
-            <cc-color-band-row mapColorFor="negative" [count]="colorCategoryCounts().negative"></cc-color-band-row>
-        } @else {
-            <cc-color-band-row mapColorFor="positiveDelta"></cc-color-band-row>
-            <cc-color-band-row mapColorFor="negativeDelta"></cc-color-band-row>
-        }
-        <cc-color-band-row mapColorFor="selected"></cc-color-band-row>
-    </div>
+    <cc-color-bands-section></cc-color-bands-section>
 
-    <div class="flex items-center justify-between gap-2">
-        @if (!isDeltaState()) {
-            <label class="flex cursor-pointer items-center gap-2">
-                <input
-                    type="checkbox"
-                    class="toggle toggle-sm toggle-primary"
-                    [checked]="isColorRangeInverted()"
-                    (change)="handleIsColorRangeInvertedChange()"
-                />
-                <span class="text-sm">Invert colors</span>
-            </label>
-        } @else {
-            <label class="flex cursor-pointer items-center gap-2">
-                <input
-                    type="checkbox"
-                    class="toggle toggle-sm toggle-primary"
-                    [checked]="areDeltaColorsInverted()"
-                    (change)="handleAreDeltaColorsInvertedChange()"
-                />
-                <span class="text-sm">Invert colors</span>
-            </label>
-        }
-        <cc-reset-settings-button
-            [settingsKeys]="resetColorsKeys()"
-            [tooltip]="isDeltaState() ? 'Reset color values to default' : 'Reset color values and gradient mode to default'"
-            label="Reset colors"
-            [small]="true"
-        ></cc-reset-settings-button>
-    </div>
+    <cc-invert-reset-row></cc-invert-reset-row>
 
     <cc-folder-overrides></cc-folder-overrides>
 </cc-settings-popover-shell>
