+++ visualization/app/codeCharta/features/metricsBar/components/colorSettingsPopover/colorSettingsPopover.component.spec.ts
@@ -1,24 +1,15 @@
-import { TestBed } from "@angular/core/testing"
 import { State } from "@ngrx/store"
-import { MockStore, provideMockStore } from "@ngrx/store/testing"
-import { render } from "@testing-library/angular"
+import { provideMockStore } from "@ngrx/store/testing"
+import { render, screen } from "@testing-library/angular"
 import { of } from "rxjs"
-import { invertColorRange, invertDeltaColors } from "../../../../state/store/appSettings/mapColors/mapColors.actions"
-import { defaultMapColors } from "../../../../state/store/appSettings/mapColors/mapColors.reducer"
-import { mapColorsSelector } from "../../../../state/store/appSettings/mapColors/mapColors.selector"
-import { setColorRange } from "../../../../state/store/dynamicSettings/colorRange/colorRange.actions"
-import { calculateInitialColorRange } from "../../../../state/store/dynamicSettings/colorRange/calculateInitialColorRange"
 import { isDeltaStateSelector } from "../../../../state/selectors/isDeltaState.selector"
+import { colorMetricSelector } from "../../../../state/store/dynamicSettings/colorMetric/colorMetric.selector"
 import { defaultState } from "../../../../state/store/state.manager"
 import { CodeMapRenderService } from "../../../../ui/codeMap/codeMap.render.service"
 import { ColorSettingsPopoverComponent } from "./colorSettingsPopover.component"
 
 describe("ColorSettingsPopoverComponent", () => {
-    afterEach(() => {
-        jest.useRealTimers()
-    })
-
-    async function setup(mapColors = defaultMapColors, isDeltaState = false) {
+    async function setup(colorMetric = "mcc", isDeltaState = false) {
         const renderResult = await render(ColorSettingsPopoverComponent, {
             inputs: {
                 popoverId: "metric-settings-popover-color",
@@ -28,7 +19,7 @@ describe("ColorSettingsPopoverComponent", () => {
                 provideMockStore({
                     initialState: defaultState,
                     selectors: [
-                        { selector: mapColorsSelector, value: mapColors },
+                        { selector: colorMetricSelector, value: colorMetric },
                         { selector: isDeltaStateSelector, value: isDeltaState }
                     ]
                 }),
@@ -43,148 +34,39 @@ describe("ColorSettingsPopoverComponent", () => {
                 }
             ]
         })
-        return { component: renderResult.fixture.componentInstance, fixture: renderResult.fixture }
+        return { component: renderResult.fixture.componentInstance }
     }
 
-    it("should dispatch setColorRange with pending values after the debounce time elapses", async () => {
-        // Arrange
-        jest.useFakeTimers()
-        const { component } = await setup()
-        const store = TestBed.inject(MockStore)
-        const dispatchSpy = jest.spyOn(store, "dispatch")
-
-        // Act
-        component.handleValueChange({ newLeftValue: 5, newRightValue: 42 })
-        jest.advanceTimersByTime(400)
-
-        // Assert
-        expect(dispatchSpy).toHaveBeenCalledWith(setColorRange({ value: { from: 5, to: 42 } }))
-    })
-
-    it("should debounce rapid changes into a single dispatch using the latest values", async () => {
-        // Arrange
-        jest.useFakeTimers()
-        const { component } = await setup()
-        const store = TestBed.inject(MockStore)
-        const dispatchSpy = jest.spyOn(store, "dispatch")
-
-        // Act
-        component.handleValueChange({ newLeftValue: 1, newRightValue: undefined })
-        jest.advanceTimersByTime(100)
-        component.handleValueChange({ newLeftValue: undefined, newRightValue: 9 })
-        jest.advanceTimersByTime(400)
-
-        // Assert
-        expect(dispatchSpy).toHaveBeenCalledTimes(1)
-        expect(dispatchSpy).toHaveBeenCalledWith(setColorRange({ value: { from: 1, to: 9 } }))
-    })
-
-    it("should flush a pending color range when destroyed before the debounce timer fires", async () => {
-        // Arrange
-        jest.useFakeTimers()
-        const { component } = await setup()
-        const store = TestBed.inject(MockStore)
-        const dispatchSpy = jest.spyOn(store, "dispatch")
-
-        // Act
-        component.handleValueChange({ newLeftValue: 5, newRightValue: 42 })
-        component.ngOnDestroy()
-
-        // Assert: the adjustment is committed instead of silently discarded
-        expect(dispatchSpy).toHaveBeenCalledWith(setColorRange({ value: { from: 5, to: 42 } }))
-
-        jest.advanceTimersByTime(400)
-        expect(dispatchSpy).toHaveBeenCalledTimes(1)
-    })
-
-    it("should report colors as not inverted for the default map colors", async () => {
-        // Arrange & Act
-        const { component } = await setup(defaultMapColors)
-
-        // Assert
-        expect(component.isColorRangeInverted()).toBe(false)
-        expect(component.areDeltaColorsInverted()).toBe(false)
-    })
-
-    it("should report colors as inverted based on the explicit inversion flag", async () => {
-        // Arrange & Act
-        const { component } = await setup({
-            ...defaultMapColors,
-            positive: defaultMapColors.negative,
-            negative: defaultMapColors.positive,
-            isColorRangeInverted: true
-        })
-
-        // Assert
-        expect(component.isColorRangeInverted()).toBe(true)
-    })
-
-    it("should keep reporting inversion when individual colors were customized afterwards", async () => {
-        // Arrange & Act
-        const { component } = await setup({
-            ...defaultMapColors,
-            positive: "#123456",
-            isColorRangeInverted: true
-        })
-
-        // Assert: the checkbox must not desync once a color is customized
-        expect(component.isColorRangeInverted()).toBe(true)
-    })
-
-    it("should report delta colors as inverted based on the explicit inversion flag", async () => {
-        // Arrange & Act
-        const { component } = await setup({
-            ...defaultMapColors,
-            positiveDelta: defaultMapColors.negativeDelta,
-            negativeDelta: defaultMapColors.positiveDelta,
-            areDeltaColorsInverted: true
-        })
-
-        // Assert
-        expect(component.areDeltaColorsInverted()).toBe(true)
-    })
-
-    it("should dispatch the recalculated initial color range when resetting the thresholds", async () => {
-        // Arrange
-        const { component } = await setup()
-        const store = TestBed.inject(MockStore)
-        const dispatchSpy = jest.spyOn(store, "dispatch")
-
-        // Act
-        component.resetColorRange()
-
-        // Assert: the header reset icon must restore the calculated thresholds
-        expect(dispatchSpy).toHaveBeenCalledWith(setColorRange({ value: calculateInitialColorRange({ minValue: 0, maxValue: 0 }) }))
-    })
-
-    it("should reset the gradient mode together with the colors outside delta mode", async () => {
+    it("should compose all sections for a regular color metric", async () => {
         // Arrange & Act
         const { component } = await setup()
 
         // Assert
-        expect(component.resetColorsKeys()).toContain("dynamicSettings.colorMode")
+        expect(component.hasRangeSection()).toBe(true)
+        expect(screen.getByText("Distribution")).not.toBeNull()
+        expect(screen.getByText("Gradient Mode")).not.toBeNull()
+        expect(screen.getByText("Bands")).not.toBeNull()
+        expect(screen.getByText("Invert colors")).not.toBeNull()
+        expect(screen.getByText("Folder Overrides")).not.toBeNull()
     })
 
-    it("should not reset the gradient mode in delta mode where it is not configurable", async () => {
+    it("should hide the range and gradient sections for the unary metric", async () => {
         // Arrange & Act
-        const { component } = await setup(defaultMapColors, true)
+        const { component } = await setup("unary")
 
         // Assert
-        expect(component.resetColorsKeys()).not.toContain("dynamicSettings.colorMode")
+        expect(component.hasRangeSection()).toBe(false)
+        expect(screen.queryByText("Distribution")).toBeNull()
+        expect(screen.queryByText("Gradient Mode")).toBeNull()
+        expect(screen.getByText("Bands")).not.toBeNull()
     })
 
-    it("should dispatch invertColorRange and invertDeltaColors from the toggle handlers", async () => {
-        // Arrange
-        const { component } = await setup()
-        const store = TestBed.inject(MockStore)
-        const dispatchSpy = jest.spyOn(store, "dispatch")
-
-        // Act
-        component.handleIsColorRangeInvertedChange()
-        component.handleAreDeltaColorsInvertedChange()
+    it("should hide the range and gradient sections in delta mode", async () => {
+        // Arrange & Act
+        const { component } = await setup("mcc", true)
 
         // Assert
-        expect(dispatchSpy).toHaveBeenCalledWith(invertColorRange())
-        expect(dispatchSpy).toHaveBeenCalledWith(invertDeltaColors())
+        expect(component.hasRangeSection()).toBe(false)
+        expect(screen.queryByText("Distribution")).toBeNull()
     })
 })
