+++ visualization/app/codeCharta/features/metricsBar/components/colorSettingsPopover/folderOverrides.component.html
@@ -10,59 +10,13 @@
 @if (overrides().length > 0) {
     <div class="flex max-h-36 flex-col gap-0.5 overflow-y-auto">
         @for (override of overrides(); track override.path) {
-            <div class="flex shrink-0 items-center gap-2 rounded-md border border-base-300 px-2 py-1" [title]="override.path">
-                <cc-inline-color-picker
-                    [hexColor]="override.color"
-                    [ariaLabel]="'Recolor ' + override.path"
-                    (colorChange)="handleRecolor(override.path, $event)"
-                ></cc-inline-color-picker>
-                <span class="min-w-0 flex-1 truncate text-sm">{{ override.path }}</span>
-                <span class="text-xs opacity-60 tabular-nums" [title]="override.fileCount + ' files'">{{ override.fileCount }}</span>
-                <button class="btn btn-ghost btn-xs btn-square" [title]="'Unpin ' + override.path" (click)="handleUnpin(override.path)">
-                    <i class="fa fa-times"></i>
-                </button>
-            </div>
+            <cc-folder-override-row
+                [override]="override"
+                (recolorOverride)="handleRecolor(override.path, $event)"
+                (unpinOverride)="handleUnpin(override.path)"
+            ></cc-folder-override-row>
         }
     </div>
 }
 
-@if (!isPinning()) {
-    <button
-        class="btn btn-ghost btn-sm w-full justify-start gap-2 border border-dashed border-base-300 font-normal"
-        (click)="startPinning()"
-    >
-        <i class="fa fa-plus"></i>
-        <span>Pin a folder color…</span>
-    </button>
-} @else {
-    <label class="input input-sm input-bordered flex items-center gap-2">
-        <i class="fa fa-search opacity-60"></i>
-        <input
-            #pinInput
-            type="text"
-            class="grow"
-            placeholder="Search folders…"
-            [value]="searchTerm()"
-            (input)="handleSearchTermChange($any($event.target).value)"
-            (keydown.escape)="handleSearchEscape($event)"
-            (blur)="handleSearchBlur()"
-        />
-    </label>
-    <ul class="m-0 flex max-h-40 list-none flex-col overflow-y-auto p-0">
-        @for (path of folderSuggestions(); track path) {
-            <li class="flex">
-                <!-- mousedown would blur the input and close the empty search before the click can pin -->
-                <button
-                    class="btn btn-ghost btn-xs w-full flex-nowrap justify-start font-normal"
-                    [title]="path"
-                    (mousedown)="$event.preventDefault()"
-                    (click)="handlePin(path)"
-                >
-                    <span class="min-w-0 truncate">{{ path }}</span>
-                </button>
-            </li>
-        } @empty {
-            <li class="px-2 py-1 text-xs opacity-60">No folders found</li>
-        }
-    </ul>
-}
+<cc-pin-folder-search (pinFolder)="handlePin($event)"></cc-pin-folder-search>
